Move flake8 options to config
This is a straight refactor of the `lint` target to move the flake8
options from the command line into a config file. This makes it easier
to document and change them, and will make it easier to centralize this
configuration across multiple projects once we've decided on a common
standard.