Port Python webhook framework to Go
This PR introduces a ground-up rewrite of the Python framework in Go.
The rewrite aims to be functionally identical to the original Python
webhook (except the rewrite uses `admissionregistration.k8s.io/v1`
instead of `v1beta1`), keeping pre-existing webhooks at the same URI.
The original home for this was my in my personal github
[lisa/k8s-webhook-framework](https://github.com/lisa/k8s-webhook-framework)
for development.

There are several major changes:

1. There are no longer YAML template files.  2. The YAML SelectorSyncSet
is created with Go (see `build/syncset.go`).  3. All webhooks must be
implemented in Go, and satisfy a specific interface (see
`pkg/webhooks/register.go` and new `README.md`).  4. App-SRE PR check
and build files have been altered.  5. The init container is rewritten
in Go and included in the container image (as
`/usr/local/bin/injector`).  6. The rewrite uses
`admissionregistration.k8s.io/v1` instead of `v1beta1`.  7. There is no
ability to make changes to the configuration by environment variables;
they are in source code (see [Method](#method)).

(For Red Hatters, refer to May 27, 2020 SREP Sprint 184 demo for
additional information)

The motivation for the rewrite centres around scaling and handling
growing pains. When the Python framework was first created, the scope of
it was quite small, handling only a single hook (Namespace validation),
focused around SREP and intended to be a stopgap measure until other
alternatives could be put into place.

Since the inception of the Python framework we have seen the scope grow
(and even shrink), and other teams contributing meaningful hooks.

Moving to Go increases our ability to add more webhooks in a consistent
way (refer to the interface in `pkg/webhooks/register.go` and
`build/syncset.go`) while leaning into our Go expertise for its inherent
type safety.

The method behind the rewrite centred around the creation of a single,
common interface to ensure a each webhook were as uniform in
implementation as possible. The benefit is two-fold. First, when
creating the webserver (`cmd/main.go`), each webhook is uniformly
handled. Second, to create the SelectorSyncSet template, the same
interface is used to iterate through all registered webhooks to
programatically create the SelectorSyncSet template, removing the need
for the previous `/templates` directory. Once the templates directory is
removed the tedium of ensuring the correct content is copied/pasted in
the right place, with the right name is eliminated, along with one more
chance for error.

The Go framework follow the "factory" and "registration" pattern where
each webhook adds a file in `pkg/webhooks` to "register" itself using
the `Register` function (`pkg/webhooks/register.go`), which in requires
a webhook name and factory function as parameters. These inputs are
stored in a `Webhooks` map (`[string)]WebhookFactory{}`) that consumers
(such as `cmd/main.go` and `build/syncset.go`) may use to access each
registered webhook.

Environment variable configuration has been removed in favour of source
code changes (or perhaps a [configuration
file](https://github.com/lisa/k8s-webhook-framework/issues/32)). The
reasoning behind removing environment variables is that changes to
environment variables require changes to the SelectorSyncSet, which
force a new build, so there is just as much work required to make a
source code change. Additionally, relying on source code change means
those changes can be run through the unit test framework to give a level
of comfort for the changes.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>