OSD-4391: Build syncset in Go container b/c pipeline lacks go
The pipeline doesn't have `go` in it (and if it did, it wouldn't be
guaranteed to be the version used herein), so build the syncset make
target in a golang:1.14 container.

Other cosmetic fixes to the Makefile (white spacing, use of $@).

It is tempting to use
`registry.svc.ci.openshift.org/openshift/release:golang-1.14` for the
syncset target since that's the base image used in the `Dockerfile`, but
in my testing that image did not work for this.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>