Add ObjectSelector support to Webhook interface
Starting in Kubernetes 1.15, webhooks gained the ability to use a
LabelSelector to "limit which requests are intercepted" based on the
labels of objects that would be sent. This PR adds to the Webhook
interface an `ObjectSelector() *metav1.LabelSelector` method which will
support this. It is recommended to set to `nil` if the particular
webhook has no such ObjectSelector restrictions (so that the resulting
selectorsyncset.yaml file does not include an empty map, for clarity
sake).

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>