OSD-6482: Dedicated-admins should be able to manage CustomDomain
This will allow `CustomDomain` objects from `managed.openshift.io` to be
managed by members of the dedicated-admins group. This will also
slightly modify the logic of the `MustGather` section to refer to the
ApiGroup as well as the Kind (for future safety), and switch them to
being references to constants. Further, the check is done outside a loop
so it's a wee bit faster.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>