OSD-7102 move to k8s.io/api/admission/v1
A number of things have changed to support this change with the most
obvious being the transition from `k8s.io/api/admission/v1beta1` to
`k8s.io/api/admission/v1`. This has required the upgrading of _many_
packages (listed in `go.mod`):

* k8s.io/api to v0.21.1
* k8s.io/apimachinery to v0.21.1
* k8s.io/client-go to v0.21.1
* sigs.k8s.io/controller-runtime to v0.8.3
* github.com/openshift/api to `e273a339932a`

Other minor changes as well to go.mod in support of those package
version bumps.

Notably is the addition of klog v2.9.0 because the logging facilities
from controller-runtime have changed. This unfortunately means the log
formatting has changed as well.

Secondary is the openshift/api package (used for syncset generation),
which has deeper requirements that were conflicting. It depends on the
k8s.io v0.21.1 version, so that is why it's used here, too.

All places in code have been updated to reflect the change to use the
`v1` API.

For development purposes I have also modified the Makefile to allow for
people to specify a `V=1` to `make` to see commands that are being run.
The default behaviour is still to hide this and act as before; sample
usage: `make V=1 test`.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>